# be/docs/swagger.yaml
definitions:
...
  internal_handlers.errorResponse:
    properties:
      code:
        example: pool_too_small
        type: string
      details: {}
      error:
        example: descriptive error message
        type: string
    type: object
...
  internal_handlers.messageResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
...
  internal_handlers.putRatingRequest:
    properties:
      item_id:
        type: string
      item_type:
        enum:
        - test
        - content
        - brain_hack
        - flashcard_deck
        - question
        type: string
      tags:
        items:
          type: string
        type: array
      value:
        type: integer
    type: object
...
info:
  contact: {}
paths:
...
  /api/v1/ratings:
    put:
      consumes:
      - application/json
      description: Stars 1-5 for tests, videos/documents, brain hacks and flashcard
        decks; thumbs (-1/+1) for a question's explanation. Only after the student
        actually did the thing (submitted the test, watched the video, saw the question)
        — else 403 `not_eligible`. No free text; `tags` come from a fixed list. Rating
        twice updates the rating.
      parameters:
      - description: Rating
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_handlers.putRatingRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_handlers.errorResponse'
      security:
      - BearerAuth: []
      summary: Rate an item (create or change)
      tags:
      - Ratings
  /api/v1/ratings/{item_type}/{item_id}:
    delete:
      parameters:
      - description: Item type
        in: path
        name: item_type
        required: true
        type: string
      - description: Item UUID
        in: path
        name: item_id
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_handlers.messageResponse'
      security:
      - BearerAuth: []
      summary: Remove my rating
      tags:
      - Ratings
...
swagger: "2.0"